requirements.txt
torch==1.11.0
torchvision==0.12.0+cu113
tqdm==4.61.2
scikit-learn==1.3.2
pandas==2.0.3
Pillow==9.1.1
wandb==0.19.11
